Extract Textarea component from Home
